Update MetalFx upscaling settings.

+++ Limelight/Stream/StreamConfiguration.h
@@ -28,7 +28,7 @@
 @property BOOL swapABXYButtons;
 @property int audioConfiguration;
 @property int supportedVideoFormats;
-@property int metalFxMultiplier;
+@property float metalFxMultiplier;
 @property BOOL multiController;
 @property BOOL useFramePacing;
 @property NSData* serverCert;
