+++ Limelight/Stream/VideoDecoderRenderer.h
@@ -17,7 +17,7 @@
 @interface VideoDecoderRenderer : NSObject
 
 - (id)initWithView:(UIView*)view callbacks:(id<ConnectionCallbacks>)callbacks streamAspectRatio:(float)aspectRatio useFramePacing:(BOOL)useFramePacing
- metalFxMultiplier:(NSInteger)metalFxMultiplier;
+ metalFxMultiplier:(float)metalFxMultiplier;
 
 - (void)setupWithVideoFormat:(int)videoFormat width:(int)videoWidth height:(int)videoHeight frameRate:(int)frameRate;
 - (void)start;
